[Gold III] Title: 두 배열의 합, Time: 1088 ms, Memory: 51028 KB -BaekjoonHub

+++ 백준/Gold/2143. 두 배열의 합/두 배열의 합.java
@@ -89,7 +89,7 @@ public class Main {
 		long count = 0;
 		
 		while(pointerA < listA.size() && pointerB < listB.size()) {
-			long sum = listA.get(pointerA) + listB.get(pointerB);
+			int sum = listA.get(pointerA) + listB.get(pointerB);
 			
 			if(sum > T) {
 				pointerB++;
